+++ front/src/Components/header/header-desktop/header-desktop-states/headerLogout.css
@@ -1,0 +1,66 @@
+.log-icon {
+    width: 3.125em;
+    height: 3.125em;
+    filter: drop-shadow(0px 4px 4px rgba(0, 0, 0, 0.25));
+    border-radius: 50%;
+    cursor: pointer;
+}
+
+.header-menu {
+    position: relative;
+    width: 3.125em;
+  }
+  
+  .logo-header-menu {
+    width: 2.875em;
+  }
+    
+  .menu-items {
+    list-style: none;
+    padding: 0;
+    margin: 0;
+    position: absolute;
+    top: 3.75em;
+    width: 12.25em;
+    right: -12.5em;
+    height: 8em;
+    box-shadow: 0 0 10px rgba(0, 0, 0, 0.2);
+    transition: 0.4s;
+    opacity: 0;
+  }
+  
+  .menu-items.open {
+    right: 0;
+    opacity: 1;
+    background-color: var(--color-bg);
+    border-radius: 10px;
+  }
+  
+  .menu-items li {
+    cursor: pointer;
+    padding: 0.625em;
+    text-align: center;
+  }
+
+  .menu-cat {
+    border-bottom: 1px solid var(--color-font-medium);
+  }
+
+  .menu-cat:hover {
+   background-color: var(--color-bg-btn-light);
+  }
+
+  .burger-menu.top:hover {
+    border-radius: 10px 10px 0 0;
+  }
+
+  
+  .menu-cat-deco {
+    background-color: var(--color-bg-btn-light);
+    border-radius: 0 0 10px 10px;
+  }
+
+  .menu-cat-deco:hover {
+    color: var(--color-bg);
+  }
+
